screen after & select dish are both in screen2.html
Just updated the way the "My dinner" table on the left of the screen
works, which is the only difference we had between the two views "screen
after" and "select dish".

If there are no selected dishes in the dinner model (id/value of the
dish is set to 0) it doesn't display the dish. If there are no dishes in
the model (every dish's id/value is 0) it just says "Pending " on the
left side of the screen.