index and screen2 in same file
added controller/stateController.js

This code probably has to change as we add more views and buttons
(something like an array of buttons or something) but right now i was
just testing.

If you now open index.html and click "Create new dinner" the screen2 div
will be shown and screen1 div will be hidden (this is done with
style="display:none;" and style="display:block;"). When you come to
screen2 and click "Confirm dinner" it does the same thing and switches
back to "screen1"